11:54

+++ include/Board.h
@@ -18,7 +18,7 @@ public:
 	std::vector<std::string> getBoard() const;
 	void updateBoard(std::vector<std::unique_ptr<MovingObjects>> & cats, 
 		             std::unique_ptr<MovingObjects > &mouse,
-					 double& levelTime);
+					 float& levelTime);
 	void updateObjects(std::vector<std::unique_ptr<MovingObjects>>& cats, std::unique_ptr<MovingObjects >& mouse);
 	void updateMembers(std::vector<std::unique_ptr<MovingObjects>>& cats, std::unique_ptr<MovingObjects >& mouse,
 		               const char character, int row, int col);
